Main

+++ Semester/check_total_no_students_and_total_no_schools_sr.py
@@ -20,11 +20,17 @@ class grade_subject_dropdowns():
         count = 0
         self.driver.find_element_by_xpath(Data.hyper_link).click()
         self.data.page_loading(self.driver)
+        year = Select(self.driver.find_element_by_id('year'))
+        self.year = year.first_selected_option.text
+        semester = self.driver.find_element_by_id('choose_semester').get_attribute('value')
+        value = semester.split(":")
+        self.semester = value[1].strip()
         grade =Select(self.driver.find_element_by_id(Data.Grade))
         counter = len(grade.options)
         for i in range(1,len(grade.options)):
             grade.select_by_index(i)
             print(grade.options[i].text)
+            time.sleep(2)
             self.data.page_loading(self.driver)
         return counter
 
@@ -34,6 +40,11 @@ class grade_subject_dropdowns():
         p = pwd()
         files = file_extention()
         self.driver.implicitly_wait(100)
+        year = Select(self.driver.find_element_by_id('year'))
+        self.year = year.first_selected_option.text
+        semester = self.driver.find_element_by_id('choose_semester').get_attribute('value')
+        value = semester.split(":")
+        self.semester = value[1].strip()
         management = self.driver.find_element_by_id('name').text
         management = management[16:].lower().strip()
         self.driver.find_element_by_xpath(Data.hyper_link).click()
@@ -50,7 +61,7 @@ class grade_subject_dropdowns():
             self.data.page_loading(self.driver)
             self.driver.find_element_by_id(Data.Download).click()
             time.sleep(4)
-            self.filename = p.get_download_dir() + '/'+ files.sr_gradewise()+management+"_all_Grade_"+gradenum.strip()+'__allDistrict_'+self.data.get_current_date()+'.csv'
+            self.filename = p.get_download_dir() + '/'+ files.sr_gradewise()+management+"_"+self.year.strip()+"_"+self.semester+"_Grade_"+gradenum.strip()+'__allDistricts_'+self.data.get_current_date()+'.csv'
             print(self.filename)
             time.sleep(1)
             if os.path.isfile(self.filename) != True:
@@ -71,31 +82,46 @@ class grade_subject_dropdowns():
         count = 0
         self.driver.find_element_by_xpath(Data.hyper_link).click()
         self.data.page_loading(self.driver)
+        year = Select(self.driver.find_element_by_id('year'))
+        self.year = year.first_selected_option.text
+        semester = self.driver.find_element_by_id('choose_semester').get_attribute('value')
+        value = semester.split(":")
+        self.semester = value[1].strip()
         management = self.driver.find_element_by_id('name').text
         management = management[16:].lower().strip()
         grade = Select(self.driver.find_element_by_id(Data.Grade))
         subjects = Select(self.driver.find_element_by_id(Data.Subject))
         subcount = len(subjects.options)-1
         files = file_extention()
-        for i in range(1,len(grade.options)):
+        for i in range(1,len(grade.options)-1):
             grade.select_by_index(i)
+            time.sleep(2)
+            print(grade.options[i].text)
             self.data.page_loading(self.driver)
             gradename = (grade.options[i].text).strip()
             gradenum = re.sub('\D','',gradename)
-            for j in range(1,len(subjects.options)):
+            for j in range(1,len(subjects.options)-1):
+                time.sleep(1)
                 subjects.select_by_index(j)
-                self.data.page_loading(self.driver)
-                sub = (subjects.options[j].text).strip()
-                self.driver.find_element_by_id(Data.Download).click()
-                time.sleep(3)
-                self.filename = p.get_download_dir() + '/' + files.sr_gradewise()+management+"_all_Grade_"+gradenum.strip()+'_'+sub+'_allDistrict_' + self.data.get_current_date()+'.csv'
-                print(self.filename)
-                if os.path.isfile(self.filename) != True:
-                    print(files.sr_gradewise()+gradenum.strip() ,' wise csv file is not downloaded')
-                    count = count + 1
+                time.sleep(2)
+                print(subjects.options[j].text)
+                if 'No data found' in self.driver.page_source:
+                    print(grade.options[i].text+', '+subjects.options[j].text+" is not having data")
+                    return count
+                else:
+                    self.data.page_loading(self.driver)
+                    sub = (subjects.options[j].text).strip()
+                    time.sleep(3)
+                    self.driver.find_element_by_id(Data.Download).click()
+                    time.sleep(3)
+                    self.filename = p.get_download_dir() + '/' + files.sr_gradewise()+management+"_"+self.year.strip()+"_"+self.semester+"_Grade_"+gradenum.strip()+'_'+sub+'_allDistrict_' + self.data.get_current_date()+'.csv'
+                    print(self.filename)
+                    if os.path.isfile(self.filename) != True:
+                        print(files.sr_gradewise()+gradenum.strip() ,' wise csv file is not downloaded')
+                        count = count + 1
 
-                os.remove(self.filename)
-                self.data.page_loading(self.driver)
-        return count
+                    os.remove(self.filename)
+                    self.data.page_loading(self.driver)
+                return count
